feat: add TransactionDetailPage component for displaying detailed transaction information
